requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0

# AI/ML dependencies
google-generativeai>=0.3.0

# Web interface dependencies
streamlit>=1.28.0

# Optional dependencies for testing and development
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Additional dependencies for potential future features
openai>=1.0.0

# Type hints and validation
typing-extensions>=4.0.0

# Date/time utilities
python-dateutil>=2.8.0

# JSON handling (built-in but explicit for clarity)
# json - built-in module

# Regular expressions (built-in but explicit for clarity)
# re - built-in module

# Operating system interface (built-in but explicit for clarity)
# os - built-in module

# System-specific parameters (built-in but explicit for clarity)
# sys - built-in module
